backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.7
psycopg[binary]>=3.1,<3.3
python-dotenv>=1.0,<2.0
celery>=5.3,<5.5
redis>=5.0,<6.0
openai>=1.0
drf-spectacular>=0.27,<0.29
